HHH-17943 - Add test for issue
